[LOAD-24493] - Add secure flag to JSESSIONID
Why ?
Starting with 2.11, NLW will be able to put the secure flag on the JSESSIONID cookie.

How ?
Helm chart will determine if NLW is behind TLS or not based on URL scheme.
If yes, the environment variable JESSIONID_SECURE_FLAG is set to `true` and NLW will put the secure flag on the cookie.
If not, the environment variable JESSIONID_SECURE_FLAG will be set to `false` and the flag is not set on the cookie.
I also added a new value (neoload.configuration.externalTlsTermination) to let user specify that NLW will be accessed through TLS layer and that TLS termination will not be handled by the ingress configuration (value `.Values.ingress.enabled` and `.Values.ingress.tls`).